Added the series title pulldown menu. Since both the subtype and series title menu's utilize the select2 widget the same, I created in pw.lookups a LookupCascadeSelect2 service with functions to implement the query and initSelection options.
One challenge with this element is that the initial value must be set to the id in order for initSelection to be called. After that point the model property has an object rather than just the id.

Tests where added for the new LookupCascadeSelect2 service and updated the bibliodataController tests.