Created new controller and route for the edit contributor page.
Currently the controller only creates a new contributor. Code to edit and existing contributor and code to save the contributor have yet to be added. Also missing is calling the affiliation service to fill in that select button.

Moved lookups to a new services directory under mypubs. Lookups will be used on the this page and perhaps others and therefore are not specific to publications.